RootRotAI 2.1 (Web Portal)/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
Pillow==10.2.0
numpy==1.21.6
pandas==1.3.5
openpyxl==3.1.2
python-multipart==0.0.9
tensorflow==2.10.0